refactor test paths to match file structure
